Refactor _setup_queue and _on_declareok into Process
